# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Classroom Supplies Usage & Replenishment Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supplies:
    get:
      operationId: listSupplies
      tags: [supplies]
      summary: List all supplies
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of supplies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supply"
    post:
      operationId: createSupply
      tags: [supplies]
      summary: Create a supply item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyInput"
      responses:
        "201":
          description: Created supply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
...
  /supplies/{id}:
    get:
      operationId: getSupply
      tags: [supplies]
      summary: Get a supply by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supply details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "404":
          description: Not found
    patch:
      operationId: updateSupply
      tags: [supplies]
      summary: Update a supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyUpdate"
      responses:
        "200":
          description: Updated supply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "404":
          description: Not found
    delete:
      operationId: deleteSupply
      tags: [supplies]
      summary: Delete a supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /supplies/{id}/restock:
    post:
      operationId: restockSupply
      tags: [supplies]
      summary: Restock a supply item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestockInput"
      responses:
        "200":
          description: Updated supply after restock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "404":
          description: Not found
  /usage-logs:
    get:
      operationId: listUsageLogs
      tags: [usage-logs]
      summary: List usage logs
      parameters:
        - name: supplyId
          in: query
          required: false
          schema:
            type: integer
        - name: classroomId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of usage logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsageLog"
    post:
      operationId: createUsageLog
      tags: [usage-logs]
      summary: Log supply usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageLogInput"
      responses:
        "201":
          description: Created usage log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageLog"
        "400":
          description: Insufficient stock
  /usage-logs/{id}:
    get:
      operationId: getUsageLog
      tags: [usage-logs]
      summary: Get a usage log by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usage log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageLog"
        "404":
          description: Not found
    delete:
      operationId: deleteUsageLog
      tags: [usage-logs]
      summary: Delete a usage log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/usage:
    get:
      operationId: getUsageReport
      tags: [reports]
      summary: Get usage report data
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
        - name: classroomId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                    quantityUsed:
                      type: number
                    supplyName:
                      type: string
                    classroomName:
                      type: string
...
components:
  schemas:
...
    Supply:
      type: object
      required: [id, name, category, quantity, unit, reorderThreshold, reorderQuantity, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reorderThreshold:
          type: number
        reorderQuantity:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SupplyInput:
      type: object
      required: [name, category, quantity, unit, reorderThreshold, reorderQuantity]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        reorderThreshold:
          type: number
          minimum: 0
        reorderQuantity:
          type: number
          minimum: 1
        notes:
          type: string
    SupplyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        reorderThreshold:
          type: number
          minimum: 0
        reorderQuantity:
          type: number
          minimum: 1
        notes:
          type: string
    RestockInput:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: number
          minimum: 1
        notes:
          type: string
    UsageLog:
      type: object
      required: [id, supplyId, supplyName, quantityUsed, usedBy, usedAt]
      properties:
        id:
          type: integer
        supplyId:
          type: integer
        supplyName:
          type: string
        classroomId:
          type: ["integer", "null"]
        classroomName:
          type: ["string", "null"]
        quantityUsed:
          type: number
        usedBy:
          type: string
        notes:
          type: ["string", "null"]
        usedAt:
          type: string
    UsageLogInput:
      type: object
      required: [supplyId, quantityUsed, usedBy]
      properties:
        supplyId:
          type: integer
        classroomId:
          type: integer
        quantityUsed:
          type: number
          minimum: 0.01
        usedBy:
          type: string
          minLength: 1
        notes:
          type: string
...